[feat] 메세지 삭제 api 예외처리 추가

+++ src/main/java/com/hotsix/server/message/service/MessageService.java
@@ -4,6 +4,7 @@ import com.hotsix.server.global.Rq.Rq;
 import com.hotsix.server.global.exception.ApplicationException;
 import com.hotsix.server.message.dto.MessageRequestDto;
 import com.hotsix.server.message.entity.Message;
+import com.hotsix.server.message.exception.MessageErrorCase;
 import com.hotsix.server.message.repository.MessageRepository;
 import com.hotsix.server.project.entity.Project;
 import com.hotsix.server.project.exception.ProjectErrorCase;
@@ -30,6 +31,17 @@ public class MessageService {
     }
 
     public void delete(long messageId) {
+        Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new ApplicationException(MessageErrorCase.MESSAGE_NOT_FOUND));
+
+        User actor = rq.getUser();
+
+        // 작성자 본인 또는 프로젝트 관계자만 삭제 가능
+        if (!message.getSender().getUserId().equals(actor.getUserId()) &&
+            !message.getProject().getClient().getUserId().equals(actor.getUserId()) &&
+            !message.getProject().getFreelancer().getUserId().equals(actor.getUserId())) {
+            throw new ApplicationException(MessageErrorCase.FORBIDDEN_DELETE);
+        }
         messageRepository.deleteById(messageId);
     }
